ethereum: Move `Mapping` and others from `graph` to `ethereum`
