+++ graph/src/data/subgraph/mod.rs
@@ -1,13 +1,8 @@
-use anyhow::{anyhow, ensure, Error};
-use ethabi::Contract;
-use futures03::{
-    future::{try_join, try_join3},
-    stream::FuturesOrdered,
-    TryStreamExt as _,
-};
+use anyhow::{anyhow, Error};
+use futures03::{future::try_join3, stream::FuturesOrdered, TryStreamExt as _};
 use itertools::Itertools;
 use lazy_static::lazy_static;
-use semver::{Version, VersionReq};
+use semver::Version;
 use serde::de;
 use serde::ser;
 use serde_yaml;
@@ -16,7 +11,7 @@ use stable_hash::prelude::*;
 use std::{collections::BTreeSet, marker::PhantomData};
 use thiserror::Error;
 use wasmparser;
-use web3::types::{Address, H256};
+use web3::types::Address;
 
 use crate::data::store::Entity;
 use crate::data::{
@@ -35,7 +30,6 @@ use crate::{
 };
 
 use crate::prelude::{impl_slog_value, q, BlockNumber, Deserialize, Serialize};
-use crate::util::ethereum::string_to_h256;
 
 use std::fmt;
 use std::ops::Deref;
@@ -62,10 +56,6 @@ lazy_static! {
         .ok()
         .map(|s| s.eq_ignore_ascii_case("true"))
         .unwrap_or(false);
-    static ref MAX_API_VERSION: Version = std::env::var("GRAPH_MAX_API_VERSION")
-        .ok()
-        .and_then(|api_version_str| Version::parse(&api_version_str).ok())
-        .unwrap_or(Version::new(0, 0, 5));
     pub static ref MAX_SPEC_VERSION: Version = std::env::var("GRAPH_MAX_SPEC_VERSION")
         .ok()
         .and_then(|api_version_str| Version::parse(&api_version_str).ok())
@@ -449,110 +439,6 @@ pub struct Source {
     pub start_block: BlockNumber,
 }
 
-#[derive(Clone, Debug, Default, Hash, Eq, PartialEq, Deserialize)]
-pub struct TemplateSource {
-    pub abi: String,
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Deserialize)]
-pub struct UnresolvedMappingABI {
-    pub name: String,
-    pub file: Link,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct MappingABI {
-    pub name: String,
-    pub contract: Contract,
-}
-
-impl UnresolvedMappingABI {
-    pub async fn resolve(
-        self,
-        resolver: &impl LinkResolver,
-        logger: &Logger,
-    ) -> Result<MappingABI, anyhow::Error> {
-        info!(
-            logger,
-            "Resolve ABI";
-            "name" => &self.name,
-            "link" => &self.file.link
-        );
-
-        let contract_bytes = resolver.cat(logger, &self.file).await?;
-        let contract = Contract::load(&*contract_bytes)?;
-        Ok(MappingABI {
-            name: self.name,
-            contract,
-        })
-    }
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Deserialize)]
-pub struct MappingBlockHandler {
-    pub handler: String,
-    pub filter: Option<BlockHandlerFilter>,
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Deserialize)]
-#[serde(tag = "kind", rename_all = "lowercase")]
-pub enum BlockHandlerFilter {
-    // Call filter will trigger on all blocks where the data source contract
-    // address has been called
-    Call,
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Deserialize)]
-pub struct MappingCallHandler {
-    pub function: String,
-    pub handler: String,
-}
-
-#[derive(Clone, Debug, Hash, Eq, PartialEq, Deserialize)]
-pub struct MappingEventHandler {
-    pub event: String,
-    pub topic0: Option<H256>,
-    pub handler: String,
-}
-
-impl MappingEventHandler {
-    pub fn topic0(&self) -> H256 {
-        self.topic0
-            .unwrap_or_else(|| string_to_h256(&self.event.replace("indexed ", "")))
-    }
-}
-
-#[derive(Clone, Debug, Default, Hash, Eq, PartialEq, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UnresolvedMapping {
-    pub kind: String,
-    pub api_version: String,
-    pub language: String,
-    pub entities: Vec<String>,
-    pub abis: Vec<UnresolvedMappingABI>,
-    #[serde(default)]
-    pub block_handlers: Vec<MappingBlockHandler>,
-    #[serde(default)]
-    pub call_handlers: Vec<MappingCallHandler>,
-    #[serde(default)]
-    pub event_handlers: Vec<MappingEventHandler>,
-    pub file: Link,
-}
-
-#[derive(Clone, Debug)]
-pub struct Mapping {
-    pub kind: String,
-    pub api_version: Version,
-    pub language: String,
-    pub entities: Vec<String>,
-    pub abis: Vec<Arc<MappingABI>>,
-    pub block_handlers: Vec<MappingBlockHandler>,
-    pub call_handlers: Vec<MappingCallHandler>,
-    pub event_handlers: Vec<MappingEventHandler>,
-    pub runtime: Arc<Vec<u8>>,
-    pub link: Link,
-}
-
 pub fn calls_host_fn(runtime: &[u8], host_fn: &str) -> anyhow::Result<bool> {
     use wasmparser::Payload;
 
@@ -568,94 +454,6 @@ pub fn calls_host_fn(runtime: &[u8], host_fn: &str) -> anyhow::Result<bool> {
     }
 
     Ok(false)
-}
-
-impl Mapping {
-    pub fn requires_archive(&self) -> anyhow::Result<bool> {
-        calls_host_fn(&self.runtime, "ethereum.call")
-    }
-
-    pub fn has_call_handler(&self) -> bool {
-        !self.call_handlers.is_empty()
-    }
-
-    pub fn has_block_handler_with_call_filter(&self) -> bool {
-        self.block_handlers
-            .iter()
-            .any(|handler| matches!(handler.filter, Some(BlockHandlerFilter::Call)))
-    }
-
-    pub fn find_abi(&self, abi_name: &str) -> Result<Arc<MappingABI>, Error> {
-        Ok(self
-            .abis
-            .iter()
-            .find(|abi| abi.name == abi_name)
-            .ok_or_else(|| anyhow!("No ABI entry with name `{}` found", abi_name))?
-            .cheap_clone())
-    }
-}
-
-impl UnresolvedMapping {
-    pub async fn resolve(
-        self,
-        resolver: &impl LinkResolver,
-        logger: &Logger,
-    ) -> Result<Mapping, anyhow::Error> {
-        let UnresolvedMapping {
-            kind,
-            api_version,
-            language,
-            entities,
-            abis,
-            block_handlers,
-            call_handlers,
-            event_handlers,
-            file: link,
-        } = self;
-
-        let api_version = Version::parse(&api_version)?;
-
-        ensure!(
-            VersionReq::parse(&format!("<= {}", *MAX_API_VERSION))
-                .unwrap()
-                .matches(&api_version),
-            "The maximum supported mapping API version of this indexer is {}, but `{}` was found",
-            *MAX_API_VERSION,
-            api_version
-        );
-
-        info!(logger, "Resolve mapping"; "link" => &link.link);
-
-        let (abis, runtime) = try_join(
-            // resolve each abi
-            abis.into_iter()
-                .map(|unresolved_abi| async {
-                    Result::<_, Error>::Ok(Arc::new(
-                        unresolved_abi.resolve(resolver, logger).await?,
-                    ))
-                })
-                .collect::<FuturesOrdered<_>>()
-                .try_collect::<Vec<_>>(),
-            async {
-                let module_bytes = resolver.cat(logger, &link).await?;
-                Ok(Arc::new(module_bytes))
-            },
-        )
-        .await?;
-
-        Ok(Mapping {
-            kind,
-            api_version,
-            language,
-            entities,
-            abis,
-            block_handlers: block_handlers.clone(),
-            call_handlers: call_handlers.clone(),
-            event_handlers: event_handlers.clone(),
-            runtime,
-            link,
-        })
-    }
 }
 
 #[derive(Clone, PartialEq, Debug)]
