+++ graph/src/util/mod.rs
@@ -1,9 +1,6 @@
 /// Utilities for working with futures.
 pub mod futures;
 
-/// Utils for working with ethereum data types
-pub mod ethereum;
-
 /// Security utilities.
 pub mod security;
